// client/src/type/role.ts
export interface RoleInfo {
    email: string|null
    infoname: string|null
    password: string|null
    phone: string|null
    sex: number|null
    sysroleName: string|null
    sysrolecode: string|null
    username: string|null
    __v: number|null
    _id: string|null
    [key:string]:string|number|null
}